# docs/swagger.yaml
definitions:
  dto.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  file.FileOutput:
    properties:
      code:
        description: 响应代码
        type: string
      cost_time:
        description: 耗时（毫秒）
        type: integer
      file_ext:
        description: 文件扩展名
        type: string
      file_name:
        description: 原始文件名
        type: string
      file_path:
        description: 文件相对路径
        type: string
      file_size:
        description: 文件大小
        type: integer
      file_url:
        description: 文件访问URL
        type: string
      message:
        description: 响应消息
        type: string
      provider:
        description: 服务提供商名称
        type: string
      request_id:
        description: 请求ID
        type: string
      saved_name:
        description: 保存的文件名
        type: string
      success:
        description: 操作是否成功
        type: boolean
      timestamp:
        description: 响应时间戳
        type: string
      upload_time:
        description: 上传时间
        type: string
    type: object
  request.CreateAdConversionRequest:
    properties:
      ad_id:
        description: 广告ID（可选，Android无真实广告ID时为空）
        type: string
      channel_code:
        description: 渠道代码
        type: string
      conversion_type:
        description: 转化类型:1=首次打开app,2=注册,3=实名,4=留资
        maximum: 4
        minimum: 1
        type: integer
      device_id:
        description: 设备ID
        type: string
      medium:
        description: 媒介类型 广告/搜索
        type: string
      platform:
        description: 平台来源 ios/android
        type: string
    required:
    - channel_code
    - conversion_type
    type: object
  request.LoginWithSmsRequest:
    properties:
      code:
        description: 验证码
        type: string
      phone:
        description: 手机号
        type: string
    required:
    - code
    - phone
    type: object
  request.RecognizeIdCardRequest:
    properties:
      body:
        description: 图片Base64编码
        type: string
      cardSide:
        description: 卡面方向：front/back
        type: string
      url:
        description: 图片URL
        type: string
    type: object
  request.SendLoginSmsRequest:
    properties:
      phone:
        description: 手机号
        type: string
    required:
    - phone
    type: object
  request.ThreeElementsVerifyRequest:
    properties:
      idcard:
        description: 身份证号
        type: string
      name:
        description: 姓名
        type: string
    required:
    - idcard
    - name
    type: object
  response.AdConversionResponse:
    properties:
      ad_id:
        description: 广告ID
        type: string
      campaign:
        description: 活动名称
        type: string
      channel_code:
        description: 渠道代码
        type: string
      channel_name:
        description: 渠道名称
        type: string
      conversion_time:
        description: 转化时间
        type: string
      conversion_type:
        description: 转化类型
        type: integer
      conversion_type_name:
        description: 转化类型名称
        type: string
      created_at:
        description: 创建时间
        type: string
      device_has_history:
        description: 设备历史提交状态：该设备是否曾经提交过任何转化记录（用于智能重复安装处理）
        type: boolean
      device_id:
        description: 设备ID
        type: string
      id:
        description: 主键ID
        type: integer
      ip:
        description: IP地址
        type: string
      is_repeat_convert:
        description: 是否重复转化
        type: boolean
      medium:
        description: 媒介类型
        type: string
      platform:
        description: 平台来源
        type: string
      source:
        description: 流量来源
        type: string
      user_id:
        description: 用户ID
        type: integer
    type: object
  response.FileUploadResult:
    properties:
      file:
        allOf:
        - $ref: '#/definitions/file.FileOutput'
        description: 文件信息
      message:
        description: 响应消息
        type: string
      success:
        description: 上传是否成功
        type: boolean
    type: object
info:
  contact: {}
paths:
  /ad-conversion/create:
    post:
      consumes:
      - application/json
      description: 记录用户的广告转化行为，包括首次打开app、注册、留资等
      parameters:
      - description: 广告转化记录请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.CreateAdConversionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.AdConversionResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: 创建失败
          schema:
            $ref: '#/definitions/dto.Response'
      summary: 创建广告转化记录
      tags:
      - 广告转化
  /auth/identity-verification:
    post:
      consumes:
      - application/json
      description: 通过姓名和身份证号，结合token中的手机号进行三要素认证
      parameters:
      - description: 认证请求参数，包含姓名和身份证号
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.ThreeElementsVerifyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 认证成功
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: 认证失败
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: 三要素认证
      tags:
      - 身份认证
  /auth/login-with-sms:
    post:
      consumes:
      - application/json
      description: 使用手机号和验证码登录，不存在账户时自动注册
      parameters:
      - description: 登录请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.LoginWithSmsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: 登录失败
          schema:
            $ref: '#/definitions/dto.Response'
      summary: 短信验证码登录
      tags:
      - 用户认证
  /auth/send-login-sms:
    post:
      consumes:
      - application/json
      description: 向指定手机号发送登录验证码
      parameters:
      - description: 发送短信请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.SendLoginSmsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 发送成功
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: 发送失败
          schema:
            $ref: '#/definitions/dto.Response'
      summary: 发送登录短信验证码
      tags:
      - 用户认证
  /file/upload/single:
    post:
      consumes:
      - multipart/form-data
      description: 上传单个文件
      parameters:
      - description: 上传的文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 上传成功
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.FileUploadResult'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: 单文件上传
      tags:
      - 文件上传
  /ocr/idcard:
    post:
      consumes:
      - application/json
      description: 使用阿里云OCR识别身份证信息
      parameters:
      - description: 识别请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.RecognizeIdCardRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 识别成功
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: 识别失败
          schema:
            $ref: '#/definitions/dto.Response'
      summary: 身份证识别
      tags:
      - OCR
swagger: "2.0"